libtsdb: Huge refactoring.
This is a pretty large monolithic change that is harder to break down
into individual commits after the fact.  It includes some bug fixes, but
the main change is that we now use proper tsdb::exception objects for
error handling and we have also refactored how series locking is done.

This is in preparation for a tsdbserver tool which needs more
fine-grained control of things like locking and knowledge of exceptions
so that it can do things like return a proper error codes to the
external client when something goes wrong.