grafana: Working basic plugin that can handle Grafana queries.
This basic plugin allows the data source to be configured (just holding
the database name at this point) and allows queries to be executed.  The
query editor simply allow entry of the measurement, series and field to
be queried.  The time range is taken as the TimeRange passed in from
Grafana and then a SELECT LIMIT operation, with the limit set to the
largest uint64_t value is executed against tsdbserver.  The resulting
data points are formatted into a Frame and handed back to Grafana for
display.

No aggregation (such as mean(), min() or max()) is done at this point
and no transformations (such as derivative()) are done either - just the
raw data.  But it works and it's fast!