grafana/tgree-simpletsdb-datasource: First viable data source.
This is a large commit that isn't worth breaking up into smaller ones.
The main changes are to add more support for querying the simple_tsdb
database (COUNT_POINTS and SUM_OP support).  We use COUNT_POINTS to
figure out how many data points would be returned from a raw SELECT and
if it is a large number then we sill now do a SUM_OP instead to
downsample it for display on the screen.  We also use the COUNT_POINTS
result to just not do any query at all if it is determined that there
are no points in the range.

With these changes, it is now possible to use simple_tsdb in real-world
scenarios, and it can sample years worth of sensor data in a reasonable
amount of time.  For some rough testing on real-world sensor data of
about 1.5 years in length comprising two sensors of roughly 900M points
each, we outperformed InfluxDB significantly.