tsdbcli/main.cc: Implement MEAN.
This computes the mean of the data points, windowed according to the
specified window.  This computes the mean of the data points in each
window; it does not compute a weighted average or do an integral, so if
data points are irregularly-spaced then this may not compute what you
think it should.  This is the same behaviour as InfluxDB.

Null points are skipped over.  InfluxDB allows you to optionally fill
null points with a value instead of skipping them; we don't support that
right now because I am lazy and don't want to implement the grammar for
it.