libtsdb/mean_op.cc: Factor sum low-level code out of tsdbcli into libtsdb.
The tsdbcli code was doing a mean() operation on the data points.  A
more generic operation is to do a sum(), while also keeping track of
non_null() as you go (which is what mean() does internally).  So, we now
provide a sum_op class in libtsdb which can do the two generic
operations and then tsdbcli does the final division to convert it into
mean().

In principle, this gives us the power to do sum() and "count non_null"
along with mean(), both in tsdbcli and any other client such as
tsdbserver which will need mean() soon.