libtsdb/select_op.cc: Support decompressing .gz field files on the fly.
At first glance, when using tsdbcli and SELECT on a very large data set,
the .gz version seems much slower.  However, tsdbcli cheats and only
shows the start and end of a data chunk on SELECT.  When using MEAN,
which touches all parts of the data set, the .gz is TWICE AS FAST as the
raw uncompressed version, with the data stored on my M1 Max Macbook Pro.

There could be a lot of factors at play.  The uncompressed data files
in my test data set are around 36G while the .gz version of the same
series is only 8.8G (most of which is the uncompressed timestamp files).
I only have 32G of RAM in my Macbook, so the uncompressed data set will
definitely not fit in the page cache.  On the other hand, the .gz
version of the series will fit in the page cache and running the MEAN
test twice back-to-back on the .gz series results in a 21s runtime and
then a 9s runtime.  Running the MEAN test twice back-to-back on the
uncompressed series results in a 38s runtime both times.

Another factor is that for uncompressed data we just mmap() it and start
hitting the mapping.  If we are doing a tsdbcli-style SELECT where we
only touch a portion of a chunk then this is a huge win.  However, if we
are doing something like MEAN which touches the entire data set then we
will take a ton of page faults to load the chunk a bit at a time from
disk which the .gz version may be able to avoid.

So, for large operations .gz is a huge win.  The only downside is that
it consumes a ton of CPU cycles compared to simply doing an mmap() of
the raw data.