libtsdb/[delete|select_op].cc: Remove related TODOs.
These TODOs have to do with the consistency of the index file in a
series.  Making index file shift operations atomic has resolved the
delete.cc issues.  The select_op.cc issues aren't actually issues at all
as long as we enforce lookups to only access the live range of
timestamps, which we already do.