libtsdb/delete.cc: Do a proper fsync_and_flush() at the end of deletion.
When we delete points from the database, we really want them to be gone.
This means we need to do a full flush on the time_first update before
returning.  On the fast path, that is the only fsync operation.  On the
slow path, we update time_first, then do a barrier, then unlink all
orphaned files, then move the new index into place and do a full flush.
In both cases, the old points are guaranteed to be deleted upon return
from tsdb::delete_points() and not at some point in the future when a
flush actually happens.