libtsdb/select_op.cc: Add support for non-mmap'd timestamp files.
When scanning over a large data set, perhaps to compute the result of a
sum op, using mmap() on the timestamp files is very inefficient because
we take thousands or hundreds of thousands of page faults to load in the
timestamps from disk.  We know we are going to touch all of the
timestamps, so doing a single read() operation to pull them all in is
far better.

In the future, it might be a good idea to use aio_read() to pull in the
timestamp file while we are decompressing gzipped field data.  We could
also queue an aio_read() after having loaded all the results for this
chunk so that when we advance to the next chunk we already have the
timestamps in memory.