libtsdb/series.cc: Avoid use of O_EXLOCK.
This is a bit tricky, because we were using O_EXLOCK to atomically
create and populate some files.  Now, we create them using a temporary
name beginning with an underscore, take the lock if necessary, populate
the file and then rename it into place.  This "atomically" makes the
populated file appear, and we still hold a lock on it.