futil/xact.h: Reimplement xact_mktemp and xact_mkdtemp.
Linux doesn't have any of the _at versions of mktemp or mkdtemp.  This
is problematic, since we don't really want to rely on the CWD pointing
at any particular place (eventually, we will support having a "root"
futil::directory for each tsdb instance, allowing multiple tsdb
instances to exist in the same process if desired).

We reimplement these both by generating a random number and then trying
to exclusively create the file or directory named using that number.  If
the create fails we just try again with a new random number.  It's hacky
and inefficient but it works.