futil/ipv4.h: Return an ipv4::socket from accept().
We were returning a generic tcp::stream from accept() because we like
abstraction.  However, when we start using SSL it will be helpful to get
the actual ipv4::socket object first which we then wrap in an SSL stream
- in order to wrap it we need to know the real file descriptor number
from the original socket and that isn't available in an abstract stream.