src/tsdbcli2: Add a new implementation of tsdbcli.
On the front end, the biggest change is the use of editline so that we
can have some sort of command history now, which makes for a much better
user experience.

On the back end, the way we parse and handle commands is completely
rewritten and much more maintainable.  We can now easily do queries with
all of the different time ranges for SELECT, MEAN or COUNT without
having to do anything in the handlers for those commands - the parsing
is completely separated out into template classes.