libtsdb/sum_op: Fix buggy sum_op so we can query outside the live data range.
This does sane things when the query is for data that falls outside of
the live data range or overlaps an edge of it.  It also does sane things
if t1 < t0.  Because of the changes in the sum_op struct, we also had to
update tsdbserver at the same time.