futil/futil.h: Allow an futil::mapping of length 0.
This avoids calling mmap() when the length is 0 (which is illegal) and
instead just leaves the address as MAP_FAILED and len as 0.  Clients
already know not to dereference a mapping of length 0, so this will
simplify many constructor idioms where we are trying to map a file which
could be empty.