libtsdb/wal: Add preliminary write-ahead log support.
This can write points into a per-series write-ahead log file.  An
iterator class is provided to support access to the entries in the
write-ahead log, whose lengths can differ from series to series.  A
query class is provided to find entries in the log corresponding to a
requested time range.

The write-ahead log does not spill over into the regular series data
yet, and select operations do not query the write-ahead log for data.
Overwrite checking is not implemented.