libtsdb/delete.cc: When deleting, increment time_first past the delete timestamp.
When the client says "delete everything up to and including T", they are
stating that this series should never hold data for a timestamp in that
range ever again (since time can only go forward for a series).  The
previous behavior was to increment time_stamp at most to time_last + 1,
which could leave a gap from (time_last, T] where someone could write in
the future even though the client has requested that all such points be
deleted.

We now increment time_first all the way to T + 1.  Any writes that hit
in that gap will now be discarded (the preceding write.cc commit takes
care of that) and the client's request will be honored.  This also
simplifies the delete code since we no longer need to look at time_last.
It also makes WAL deletions trivial since all we need to do in that case
is increment time_first to T + 1.