libtsdb/select_op.cc: Refactor select_op API so that iteration ends when npoints == 0.
The existing select_op API was a bit insane when it comes to iteration.
The right way to do it is to end iteration with a 0-length chunk instead
of weird "is_last" flags and having to check that flag before advancing
if it isn't the end.  I.e. it is really weird to set "is_last" on a
result iteration that ALSO returns data.