libtsdb/series.h: Change locking so that a read lock also acquires a shared lock on time_last.
When the WAL commits to the main data store, this truncates the WAL file
down to length 0.  We don't want that to happen while we are doing a
read operation on the series.  A write operation takes an exclusive lock
on time_last, so by taking a shared lock on time_last for a read
operation we prevent the writer from possibly truncating the WAL out
from under us.