libtsdb/write.h: Make it so write_chunk_index can be mutable.
This is a bit of a hack.  Sometimes we want to use write_chunk_index to
build up a data chunk, in which case the data we are mapping should be
non-const.  Other time we want to use write_chunk_index to map data that
we received from somewhere else, in which case it should be const.

Sometimes we want to build up a write chunk using a write_chunk_index
(in tsdb::commit_wal()), where it should be mutable, but then pass it to
a function that only needs const data pointers instead (in
tsdb::write_series()).  It is really hard to make that work, so we throw
away the constness with a separate constructor here.  We may revisit
this in the future.