libtsdb/wal.cc: Commit to the main data store when the WAL overflows.
The WAL is now the only entry point for writing new points into the
database.  As the WAL fills up, it will eventually commit data into the
main data store using the tsdb::write_series() method that we have used
exclusively in the past.

The benefit of the WAL is that only a single fsync() operation is
required when a batch of points comes in.  We can repeatedly write
batches of points (or even just single points) to the WAL and not pay
the cost of writing to all of the bitmap, timestamp and field files
which require many fsync() operations and are hard on SSD longevity.

Once the size of the WAL file passes a threshold, we commit all of its
points into the main data store, which requires writing to all of those
files, however we can do it in a single large batch of points which
massively reduces the number of fsync() operations required and thus the
number of flash write operations.