libtsdb/measurement.h: Replace gen_indices() with gen_entries().
This generates a lookup table of schema_entries based on the passed-in
field names, similar to how we were generating an index list before.
This is more generic and can replace gen_indices because the index is
now embedded inside the schema_entry in the schema file.  It avoids
copying large 128-byte schema_entry objects, although the added pointer
indirection might make things a bit less efficient - hard to say.

This does allows us to completely remove the extra lookup table in
sum_op.