simple_tsdb/tsdb_client.py: Handle reconnecting the client if the connection is lost.
This also purposefully does not initiate the connection when the Client
is created, but instead defers it until the first database access.  This
allows users of the library to just wrap their database writing code in
a retry loop rather than also having to deal with Client connection
corner cases in their __init__ methods, for instance.