simple_tsdb/tsdb_client.py: Use SSL if and only if credentials are supplied.
tsdbserver requires authentication when it is running as an SSL server.
On the other hand, when tsdbserver is in non-SSL mode it is by
definition bound to localhost only, SSL is not enabled since we can't
verify certificates against internal addresses and authentication is
disabled.

In the future, we may force authentication for the loopback case even
though we aren't using SSL, but for now we can't have an instance of
use_ssl being truthy and credentials being non-truthy.