libtsdb/write.cc: Don't bother locking the index file.
The index file was only being locked in write_series(), so it could only
ever protect from access by another writer to the same series.  However,
we already hold a write_lock on the series at this point, so no other
writer can be active while we may be reading or updating the index.  A
delete operation also accesses the index file, however it holds an
exclusive lock on time_first which would prevent us from getting our
write_lock as well.