grafana: Add a Zoom option and move min/max into it, along with mean.
The min/max transform is meant to show the noise in the data set without
having to render the fine details of the data set.  That is, it is an
algorithm to decimate a large data set into something manageable for
Grafana to display - exactly the same thing as "group by" and taking the
mean of the data set.

We now separate the formatting of the data into two phases: in the first
phase we decimate using the selected Zoom option, and then in the second
phase we do the selected Transform.  So, for instance, we can now do a
Tare and display either the mean or the min/max noise for a nice
relative comparison between sensors.