simple_tsdb/client.py: Provide a mapping of simple_tsdb types to InfluxDB types.
InfluxDB seems to try and convert types like np.uint32 into float()
instead of int().  This is not very convenient, so we provide a mapping
of types here that will convert them appropriately for InfluxDB.  This
can be used to convert points read from a simple_tsdb database into a
format suitable for writing to an InfluxDB database.