tsdbserver/reflector: Add tsdbreflector binary to reflect to a remote server.
A tsdbreflector process performs forward-or-store operations on incoming
data points.  The typical use case is for a data logger to accept points
from various sensors and to try and forward them to a remote "main"
server in the cloud in real time.  Perhaps this cloud server allows
things like inspecting data points with Grafana.

At the same time, the connection to the cloud may not always be
reliable, in which case instead of backing up points in the sensor
process' RAM we want to transparently persist them in a local
simple_tsdb database until the cloud server becomes available again.
That way we can have a fairly strong guarantee that local points will
eventually make it to the safe cloud storage while at the same time
causing zero SSD wear for the common case that the cloud is reachable.