libtsdb/root.h: Make a root constructor explicit to avoid unexpected bugs.
It was possible to do this:

    tsdb::root* root = ...;
    tsdb::database db(root,database_name);

when the correct code would have been:

    tsdb::database db(*root,database_name);

This is because the root* was being converted to a bool implicitly and a
new root temporary was constructed and then passed as the root& argument
to tsdb::database().  A very wild bug to track down.

By marking the bool constructor explicit, the erroneous implicit case is
now flagged at compilation time.