grafana/tgree-simpletsdb-datasource: Add support for (and require) SSL and authentication.
This change adds hostname/username/password fields to the data source
configuration editor and then uses them to connect via SSL (instead of
plain TCP) and then authenticate with the database.  This is required so
that we can connect to the same simple_tsdb instance that is receiving
data from remote data sources such as sensors.

If Grafana and the simple_tsdb instance are running on the same server,
then the server should have an entry in its /etc/hosts file to loopback
to itself for its own hostname.