plugin/datasource.go: Don't return all-nil data frames to Grafana.
If we do a query which returns data points, but all the data points are
nil (maybe we have an uncalibrated sensor so it just returns frequencies
but not converted values and our panel is trying to plot converted
values), don't return an all-nil data frame to Grafana.  If we return
the all-nil frame then we end up with a series name in the plot's legend
but no visible data for that series, which is rather annoying.