libtsdb/wal.h: Add cast_field<> method to wal_entry.
The wal_entry doesn't know its underlying type, so it needs help for the
cast.  The type is store in the tsdb::measurement, which can be accessed
via the wal_query's series_read_lock, but we don't currently store
back-refs to the wal_query in the wal_iterator (which is where we could
move the cast to).

Maybe in the future we will update wal_iterator to know about the
wal_query and put the cast there to remove the requirement that the
caller pass in the underlying type.