plugin/datasource.go: Make min/max graphing more performant.
This was taking 10x as many data points as there were pixels in the
graph, in an attempt to make the filled-in part of the graph look
better.  In fact, this did nothing at all for the display and it looks
exactly the same if you don't take extra buckets.  This helps a lot with
the download speed on slow connections, as well as probably making the
canvas much easier to render.