tsdbserver/server.cc: Enable keepalive on server sockets.
Client sockets already use keepalive and it is needed on the server as
well in order to prevent deadlock.  What can happen without it is that a
client socket can begin a WRITE operation to the server, but then maybe
the internet goes out for a little bit.  The client's keepalive triggers
and it drops the connection to the server.  However, the server never
finds out about this (maybe the client stops sending FIN or whatever it
is that kills the TCP connection after awhile) and leaves its WRITE
operation pending for the next chunk of data from the client.

Of course, this means that the orphaned WRITE is holding filesystem
locks.  Those locks never get released and eventually we deadlock when a
new request for the same series comes in.