plugin/datasource.go: Include the first maximum value instead of skipping it for the mean value.
We make the mean of the first bucket the first data point in a min/max
series.  This is to try and "level off" the tare value between parallel
series so that they start off closer to their zero value even if they
are noisy.

If the first data point of the series is an outlier, then taking the
mean as the first point instead of the max as the first point will
truncate that spike in the initial rendered bucket.  When zoomed out on
a series, this can mislead the viewer into thinking that the series does
not start out with a large max value.

This commit inserts a max after the initial mean point so that we can
see the full range of the initial bucket while still preserving the mean
offset for inter-series taring.