tsdbserver/reflector.cc: Don't validate remote host when parsing config.
The network may not be available for some reason when tsdbreflector
starts.  This would prevent us from resolving the remote host and then
tsdbreflector would exit immediately.  This is definitely not what we
want, especially if tsdbreflector is a boot service that can launch
possibly before the network has come up!