tsdbserver/server.cc: Push the socket down into the request_handler so that it can be modified there.
Currently, we are wrapping the SSL socket before calling the handler
thread.  A malicious or buggy client can start a connection but then not
send any data ever, which hangs the wrap() call and prevents the server
from handling other new connections.  The first step to fixing this is
to push the wrap() call down into the handler thread so that the server
thread doesn't have to any IO on the new socket and so new connections
will still be able to open.  It doesn't solve the problem of zombie
connections piling up, however.