libtsdb/series.h: Open WAL file as part of series locking.
In order to obtain a file descriptor reference on the proper version of
the WAL file in a future commit, we need to open the file in the right
order between acquiring different locks.  While not necessary just now,
we move the WAL file descriptor into the lock object in preparation.