futil/futil.h: Big refactor to pull wrapper functions into futil:: namespace.
We have a lot of wrapper functions for standard unix functions.  Some of
them were embedded in classes; some of them used class references as
arguments before massaging them to file descriptors.

We hoist all of the unix wrappers out of any class definitions and into
the futil:: namespace, and we also replace any class arguments with file
descriptor numbers so that the wrapper parameters exacty match the
functions they are wrapping.

This will enable faking or mocking later, and it will also allow us to
move the wrappers into a separate header and be independent of the
utility classes definted by futil.