futil/futil_wrap.h: Add strace-like capability.
macOS doesn't have a usable strace equivalent, so it can be hard to
follow function calls and catch error conditions.  Since we are already
wrapping everything it is pretty easy for us to add something that works
a lot like strace - at least for the file I/O functions we are wrapping.