fakefs/fakefs.cc: Allow growing an mmap'd file if the vector has capacity.
This is kind of yucky, but it does allow us to test things like one
process writing to a file to make it larger while it has been mmap'd in
a different process which is working with a shorter version of the file.
File growth in practice doesn't invalidate an mmap address region; only
truncation is a problem.