libtsdb/measurement.cc: Handle create_measurement() if the measurement directory exists and is corrupt.
If the measurement directory exists, but contains no schema file, then
create_measurement() would get into an infinite retry loop because
tsdb::measurement() would continually throw
no_such_measurement_exception (because it would get ENOENT when trying
to open the schema).  The ENOENT handling was meant to detect the case
where the measurement directory doesn't exist, but because the catch
scope is too big when doing a constructor we can't really disambiguate.
This is a workaround for now, maybe we can improve it in the future.