libtsdb/root: Precompute the maximum gzipped chunk size.
This is the maximum size of a gzipped chunk; it applies only to fields
that have 8-byte entries (f64, u64).  Fields wil smaller entry sizes
will have a smaller maximum chunk size, but typically what we do is
allocate a single buffer that is reused for all fields in turn, so going
with the maximum is fine.