libtsdb/unittest: Add series test.
Note that this found that we forgot to call fsync() on a new series'
"fields" and "bitmaps" subdirectories.  Inside of those subdirectories
are a bunch of sub-subdirectories, one for each field.  Without the
fsync, a crash at the wrong time could leave us with "fields" and
"bitmaps" subdirectories but not populated with any sub-subdirectories.