libtsdb/write.cc: Add an extra scope to close an mmap() region earlier.
This closes the mmap() region before truncating the file that is being
mmapped.  fakefs doesn't like it when you truncate an mmapped file
because it isn't able to track the validity of the address space
anymore.  We don't access the region again before the truncate, so just
close it a bit earlier.