libtsdb/select_op.cc: Use in-memory gzip decompression.
This manually sets up the zstream object and does the decompression
after using futil file I/O to read it all into memory.  Doing it this
way instead of using zlib's gzFile abstraction allows us to ensure all
the file I/O is done through futil, which we will need for unittesting
in the future.