libtsdb/write.cc: Remove a set of scoping braces during failure recovery.
During failure recovery we were doing an mmap() of a tail_fd and then
later truncating it after we used the mmap to find the parts of the file
we wanted to keep.  This made fakefs angry since it isn't a good idea to
truncate while a file is mapped.

Since we no longer do a truncate but now do an atomic replace, we can
get rid of the extra scoping braces and make the code a bit more
readable.  The mmap region is now disposed after the swap happens but
since it isn't used again this is of no concern.