libtsdb/delete.cc: Fix tsdb::delete_points() so that it cleans up the result of an interrupted delete.
An interrupted delete could leave index entries at the start of the
index that don't point at anything.  They don't harm anything because
they are excluded by time_first, however they do take up space and
certainly increase complexity.  Instead of bailing out early from
delete_points() if the target timestamp is before time_first, we now
always run to completion.  As part of deletion, we now check the front
of the index file to find where the first live index entry actually is.
Normally it will just be the first entry of the file and this will take
almost no extra time.  In the case of an interrupted delete, it will
skip over all the invalid entries and later at the end of the function
we will delete them all and atomically shift the index down.