unittest/delete_test.cc: Add test propagating bad index entry in the index file.
If we do a delete operation that deletes an entire series, and crash at
the wrong time, we can be left with an entry in the index file that
points at a file that no longer exists.  A later write_points()
operation will skip past that bad entry and append a new timestamp file
to the index file after the bad entry, leaving the bad entry there
forever.  This is bad.