unittest/write_test.cc: Add test demonstrating that write_points() can truncate select_op's mmap'd region.
If a write_points() operation crashed, and left index entries at the end
of the index file that aren't covered by time_last yet due, and then we
call write_points() again, during failure recovery write_points() will
attempt to truncate the extra index entries out of the index file.

Meanwhile, select_op is perfectly allowed to have the index file mmap'd
and being doing a search on it.  The invalid entries at the end are
excluded by time_last, but index file searches use a binary search and
start from the end.  This leads to a case where select_op could be
looking at index entries that are getting truncated by write_points(),
in the best case leading to a SEGV and in the worst case causing random
behaviour.