libtsdb/write.cc: Replace tail_fd truncate with atomic-truncate.
This one is really hard to unittest so for now we just apply the fix.
Both select_op and count_comitted_points() mmap timestamp files to
search them for start and end positions.  It's imperative that the files
not be truncated during the search.  write_points() does need to shrink
the file before it writes new timestamps over the end, so we do a
file-swap so as not to invalidate the mmap'd regions (and also to not
suddenly populate part of the mmap'd regions with new timestamps,
ruining any ongoing binary search!).