libtsdb/write.cc: Massively simplify writes by detecting an empty series.
We had a number of different edge cases and were sometimes dealing with
trying to open tail files that don't even exist.  All of these are
symptoms of trying to access points in an empty series and can be easily
handled early on if we just delete every index entry if the series is
empty.