libtsdb/count.cc: Fix issue where count would return incorrect time_first value.
If the start of the range to be counted hit in the gap at the end of a
chunk file, we would return an incorrect time_first value (and
dereference an iterator past the end of the timestamp file!).

Like with the similar issue in select_op, we need to advance to the next
index slot, if it exists, or return an empty result otherwise.