libtsdb/write.cc: Populate index file before creating timestamp/field/bitmap files.
This adds a record of the files that we are going to create so that if
we crash, during failure recovery we will clean them up.  Before this
commit, we updated the index file last which meant that the
timestamp/field/bitmap files would be "lost" if a crash occurred before
the index file was written.

Having entries at the end of the index file that may not actually have
physical files backing them is OK as long as nobody tries to look at
index entries that are excluded by time_last.  With the whole series of
unittests that we've written up to this point, we are pretty confident
that the time_last setting is always honored and so it should be safe to
make this change.