Make: Don't link statically against libssl and libcrypto.
In Ubuntu 24.10, statically linking against libcrypto needs you to pull
in even more stuff, apparently from zlib (which is different from
zlib-ng just to make things more confusing).

It turns out all we need to do on Linux is specify -lssl and -lcrypto in
the linker command.  On macOS, all we need to do is point the linker at
the SSL libraries in Homebrew and then also just specify -lssl and
-lcrypto.  No need to do any static linking at all.

This also has the desirable side-effect of allowing us to build on
aarch64-linux-gnu since we were hard-coding library symlinks via
x86_64-linux-gnu before.