tmock/tmock.h: Pull TASSERT expr evaluation back up into macro.
We pushed this down into a _tassert() function in the hopes that when
debugging we would be able to see variables in higher stack frames more
easily.  This was not the case, and pushing the evaluation of expr down
hid it from the compiler and made it so it couldn't use the test for
inferences about code violations after the TASSERT that wouldn't ever be
able to actually run.