libtsdb/integral_op.cc: Fix bug accessing select_op field.
We were using the absolute schema index for the field to access
op.fields[], however op.fields[] is just the subset of fields that the
select_op was querying for.  So if we were only querying for 1 field,
but that field was at offset 3 in the schema, then it would try to
access the third element in op.fields[] which is out-of-bounds.